mooclet_engine/requirements_latest_version.txt
amqp==2.5.1
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
Click==7.0
clickclick==1.2.2
Django==2.2.28
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-cors-headers==3.1.1
django-cors-middleware==1.4.0
django-debug-toolbar==2.0
django-filter==2.2.0
django-sslserver==0.21
django-storages==1.7.2
django-timezone-field==3.1
djangorestframework==3.10.3
drf-writable-nested==0.5.1
environment==1.0.0
environments==1.0.1
Flask==1.1.1
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==2.6.0
kombu==4.6.5
MarkupSafe==1.1.1
more-itertools==7.2.0
numpy==1.21.0
openapi-spec-validator==0.2.8
pandas
python-crontab==2.3.9
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.4
requests==2.22.0
rest-pandas==1.1.0
six==1.12.0
specific==0.0.1
sqlparse==0.3.0
swagger-ui-bundle==0.0.5
the==0.1.5
to==0.3
urllib3==1.25.6
vine==1.3.0
Werkzeug==0.16.0
scipy==1.10.1
xlsxwriter
rpy2==3.5.4